Fix cache variables and move to experiments
The MPI_CMD cache variable was written as a string constant, which
was trying to run e.g. "aprun -j1" as its own command. We remove
quotation marks here to fix this. Additionally, we move NPROC
definitions for experiments to their respectives CMakeLists so
we don't have to define them at the toplevel.