Override DT for lock_exchange DX1 experiments
We can't run at the large 1800s timestep for the 1km horizontal
resolution experiments. In fact, to get stability across the
whole range of viscosities, we have to drop the timestep to 300s!
This is a bit unfortunate, because it only really affects the
initial baroclinic instability at the lowest viscosity.