lock_exchange parameter tweaks
We using NONE for buoyancy forcing instead of zero, to ensure that we
have absolutely zero forcing. We revert to a low order interpolation
scheme (there are conditioning issues with vanished layers).

The absolute biggest change here however is to use the Gudonov KE
discretisation scheme. This converts from a second-order
centred (conservative) scheme to a first-order upwind scheme. This has
the effect of imposing a scale-selective dissipation of momentum,
which is exactly what we need to reduce spurious mixing.